+++ src/pages/news/[id].tsx
@@ -8,7 +8,7 @@ import { format } from 'date-fns';
 
 import { News } from '~/interfaces/news';
 import { useLocale } from '~/hooks/useLocale';
-import { WebevOgpHead } from '~/components/commons/WebevOgpHead';
+import { WebevOgpHead } from '~/components/common/WebevOgpHead';
 
 type Props = {
   news: News;
