+++ src/pages/read/index.tsx
@@ -7,9 +7,9 @@ import { useLocale } from '~/hooks/useLocale';
 
 import { PageList } from '~/components/domain/Page/molecules/PageList';
 import { LoginRequiredWrapper } from '~/components/Authentication/LoginRequiredWrapper';
-import { SortButtonGroup } from '~/components/commons/SortButtonGroup';
-import { SearchForm } from '~/components/commons/SearchForm';
-import { WebevOgpHead } from '~/components/commons/WebevOgpHead';
+import { SortButtonGroup } from '~/components/common/SortButtonGroup';
+import { SearchForm } from '~/components/common/SearchForm';
+import { WebevOgpHead } from '~/components/common/WebevOgpHead';
 
 import { PageStatus } from '~/domains/Page';
